fix(factory): 팩토리 메서드 패턴 자바 테스트 코드 디렉토리 분리
